Uppdaterade kod med nya lösningar från parprogrammeringen

+++ LoanService/Program.cs
@@ -84,34 +84,38 @@ app.MapPost("/loan/{bookId}", async (string bookId, LoanContext ctx, HttpContext
     return Results.Created($"/loan/{loan.Id}", "Book has been loaned");
 });
 
-app.MapGet("/loans", async (LoanContext ctx) =>
+app.MapGet("/loans", async (LoanContext ctx, HttpContext httpContext) =>
 {
+    var userId = httpContext.User.Claims.FirstOrDefault(claim => claim.Type == ClaimTypes.NameIdentifier).Value;
 
-    var loans = await ctx.Loans.ToListAsync();
-    return loans;
+    if (userId == null)
+    {
+        return Results.BadRequest("Bad Token");
+    }
 
+    var loans = await ctx.Loans.Where(l => l.UserId.Equals(userId)).ToListAsync();
+    return Results.Ok(loans);
 });
 
-app.MapPost("/return/{id}", async (string id, LoanContext ctx) =>
+app.MapPost("/return/{id}", async (string id, LoanContext ctx, HttpContext httpContext) =>
 {
+    var userId = httpContext.User.Claims.FirstOrDefault(claim => claim.Type == ClaimTypes.NameIdentifier).Value;
 
-    var loan = await ctx.Loans.FirstOrDefaultAsync(u => u.BookId == id);
+    if (userId == null)
+    {
+        return Results.BadRequest("Bad Token");
+    }
+    Console.WriteLine("USERID: " + userId);
+    var loan = await ctx.Loans.Where(l => l.UserId.Equals(userId) && l.BookId.Equals(id)).FirstOrDefaultAsync();
 
     if (loan == null) return Results.NotFound("You have not loaned this book.");
 
     ctx.Loans.Remove(loan);
-    // using var channel = GrpcChannel.ForAddress("http://libraryservice");
-    // var client = new GetBookService.GetBookServiceClient(channel);
-    // var book = await client.GetBookAsync(bookRequest);
-
-    // await book.Loans.AddAsync(loan);
 
     await ctx.SaveChangesAsync();
 
     return Results.Ok("Book has been returned");
 });
-
-
 
 app.Run();
